api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
onnxruntime==1.20.1
Pillow==11.1.0
numpy==1.26.4
pydantic==2.10.4
python-multipart==0.0.20
pyyaml==6.0.2
